+++ pyproject.toml
@@ -133,6 +133,7 @@ extra-standard-library = [
     "_dummy_threading",
     "_heapq",
     "_imp",
+    "_io",
     "_json",
     "_locale",
     "_lsprof",
